+++ src/main/kotlin/dev/turingcomplete/intellijdevelopertoolsplugins/_internal/tool/generator/LoremIpsumGenerator.kt
@@ -15,8 +15,8 @@ import com.intellij.ui.dsl.builder.columns
 import com.intellij.ui.layout.ComboBoxPredicate
 import dev.turingcomplete.intellijdevelopertoolsplugins.DeveloperToolConfiguration
 import dev.turingcomplete.intellijdevelopertoolsplugins.DeveloperToolContext
-import dev.turingcomplete.intellijdevelopertoolsplugins.DeveloperToolPresentation
 import dev.turingcomplete.intellijdevelopertoolsplugins.DeveloperToolFactory
+import dev.turingcomplete.intellijdevelopertoolsplugins.DeveloperToolPresentation
 import dev.turingcomplete.intellijdevelopertoolsplugins._internal.common.ValidateMinIntValueSide.MAX
 import dev.turingcomplete.intellijdevelopertoolsplugins._internal.common.ValidateMinIntValueSide.MIN
 import dev.turingcomplete.intellijdevelopertoolsplugins._internal.common.bindIntTextImproved
@@ -29,11 +29,17 @@ import java.security.SecureRandom
 import kotlin.math.max
 import kotlin.math.min
 
-class LoremIpsumGenerator(configuration: DeveloperToolConfiguration, parentDisposable: Disposable) :
-  MultiLineTextGenerator(
+class LoremIpsumGenerator(
+  project: Project?,
+  context: DeveloperToolContext,
+  configuration: DeveloperToolConfiguration,
+  parentDisposable: Disposable
+) : MultiLineTextGenerator(
     generatedTextTitle = "Generated lorem ipsum",
     configuration = configuration,
-    parentDisposable = parentDisposable
+    parentDisposable = parentDisposable,
+    context = context,
+    project = project
   ) {
   // -- Properties -------------------------------------------------------------------------------------------------- //
 
@@ -247,7 +253,7 @@ class LoremIpsumGenerator(configuration: DeveloperToolConfiguration, parentDispo
       parentDisposable: Disposable,
       context: DeveloperToolContext
     ): ((DeveloperToolConfiguration) -> LoremIpsumGenerator) = { configuration ->
-      LoremIpsumGenerator(configuration, parentDisposable)
+      LoremIpsumGenerator(project, context, configuration, parentDisposable)
     }
   }