+++ src/main/kotlin/dev/turingcomplete/intellijdevelopertoolsplugins/_internal/tool/transformer/JsonPathTransformer.kt
@@ -3,6 +3,8 @@
 package dev.turingcomplete.intellijdevelopertoolsplugins._internal.tool.transformer
 
 import com.fasterxml.jackson.databind.node.ArrayNode
+import com.intellij.codeInsight.actions.RearrangeCodeProcessor
+import com.intellij.codeInsight.actions.ReformatCodeProcessor
 import com.intellij.icons.AllIcons
 import com.intellij.json.JsonLanguage
 import com.intellij.openapi.Disposable
@@ -13,6 +15,8 @@ import com.intellij.openapi.project.DumbAwareAction
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.ui.popup.Balloon
 import com.intellij.openapi.ui.popup.JBPopupFactory
+import com.intellij.psi.PsiManager
+import com.intellij.testFramework.LightVirtualFile
 import com.intellij.ui.ScrollPaneFactory
 import com.intellij.ui.awt.RelativePoint
 import com.intellij.ui.components.JBLabel
@@ -20,7 +24,10 @@ import com.intellij.ui.dsl.builder.Align
 import com.intellij.ui.dsl.builder.LabelPosition
 import com.intellij.ui.dsl.builder.Panel
 import com.intellij.ui.dsl.builder.RightGap
+import com.intellij.ui.dsl.builder.Row
+import com.intellij.ui.dsl.builder.bindSelected
 import com.intellij.ui.dsl.builder.bindText
+import com.intellij.ui.dsl.builder.whenStateChangedFromUi
 import com.intellij.ui.dsl.builder.whenTextChangedFromUi
 import com.intellij.util.ui.UIUtil
 import com.jayway.jsonpath.Configuration
@@ -29,6 +36,7 @@ import com.jayway.jsonpath.JsonPathException
 import com.jayway.jsonpath.spi.json.JacksonJsonNodeJsonProvider
 import com.jayway.jsonpath.spi.mapper.JacksonMappingProvider
 import dev.turingcomplete.intellijdevelopertoolsplugins.DeveloperToolConfiguration
+import dev.turingcomplete.intellijdevelopertoolsplugins.DeveloperToolConfiguration.PropertyType.CONFIGURATION
 import dev.turingcomplete.intellijdevelopertoolsplugins.DeveloperToolConfiguration.PropertyType.INPUT
 import dev.turingcomplete.intellijdevelopertoolsplugins.DeveloperToolContext
 import dev.turingcomplete.intellijdevelopertoolsplugins.DeveloperToolFactory
@@ -49,7 +57,8 @@ class JsonPathTransformer(
     sourceTitle = "Original",
     resultTitle = "Result",
     initialSourceExampleText = EXAMPLE_SOURCE,
-    initialLanguage = JsonLanguage.INSTANCE
+    inputInitialLanguage = JsonLanguage.INSTANCE,
+    outputInitialLanguage = JsonLanguage.INSTANCE
   ),
   context = context,
   configuration = configuration,
@@ -59,6 +68,7 @@ class JsonPathTransformer(
   // -- Properties -------------------------------------------------------------------------------------------------- //
 
   private val queryText = configuration.register("contentText", "", INPUT, EXAMPLE_QUERY)
+  private val formatResult = configuration.register("formatResult", true, CONFIGURATION)
   private var errorHolder = ErrorHolder()
 
   // -- Initialization ---------------------------------------------------------------------------------------------- //
@@ -73,11 +83,17 @@ class JsonPathTransformer(
         .align(Align.FILL)
         .resizableColumn()
         .gap(RightGap.SMALL)
-        .whenTextChangedFromUi { configurationChanged(QUERY_TEXT_PROPERTY_KEY, queryText) }
+        .whenTextChangedFromUi { configurationChanged(queryText) }
       lateinit var helpButton: JComponent
       helpButton = actionButton(ShowOperatorsHelpPopup { helpButton })
         .component
     }
+  }
+
+  override fun Row.buildAdditionalActionsUi() {
+    checkBox("Format result")
+      .bindSelected(formatResult)
+      .whenStateChangedFromUi { configurationChanged(queryText) }
   }
 
   override fun transform() {
@@ -89,13 +105,23 @@ class JsonPathTransformer(
     }
 
     try {
-      val result = JsonPath.parse(sourceText.get(), jsonPathConfiguration).read<Any>(query)
-      resultText.set(
-        when (result) {
-          is ArrayNode -> objectMapper.writeValueAsString(result)
-          else -> result.toString()
-        }
-      )
+      val resultJsonText = when (val resultJsonNode = JsonPath.parse(sourceText.get(), jsonPathConfiguration).read<Any>(query)) {
+        is ArrayNode -> objectMapper.writeValueAsString(resultJsonNode)
+        else -> resultJsonNode.toString()
+      }
+      if (formatResult.get()) {
+        val workingVirtualFile = LightVirtualFile(this.javaClass.canonicalName, JsonLanguage.INSTANCE, resultJsonText)
+        PsiManager.getInstance(project!!).findFile(workingVirtualFile)?.let { workingPsiFile ->
+          val processor = RearrangeCodeProcessor(ReformatCodeProcessor(project, workingPsiFile, null, false))
+          processor.setPostRunnable {
+            resultText.set(workingPsiFile.text)
+          }
+          processor.run()
+        } ?: error("snh: Can't get PSI file for `LightVirtualFile`")
+      }
+      else {
+        resultText.set(resultJsonText)
+      }
     } catch (e: JsonPathException) {
       errorHolder.add(e)
     }
@@ -194,7 +220,6 @@ class JsonPathTransformer(
 }"""
 
     private const val EXAMPLE_QUERY = "\$.starWars.characters..forename"
-    private const val QUERY_TEXT_PROPERTY_KEY = "contentText"
 
     private val operatorsHelpPanel = JBLabel(
       """